Arreglando funcionalidad

+++ borrar.php
@@ -1,17 +1,18 @@
 <?php
 include 'conexion.php';
 
- $cedula = $_REQUEST['cedula'];
+   $cedula = $_REQUEST['cedula'];
 
- $del = $con -> query("DELETE FROM usuarios WHERE cedula ='$cedula' ");
- if ($del) {
-    echo "<script>
-    location.href='formulario2.php';
-    </script>";
- }else{
-    echo "<script>
-    alert('El registro no pudo ser eliminado');
-    location.href='formulario2.php';
-    </script>";
- }
+   $sel = $con -> query("SELECT * FROM usuarios WHERE cedula='$cedula' ");
+   if($reg = $sel -> fetch_assoc()){
+      $del = $con -> query("DELETE FROM usuarios WHERE cedula ='$cedula' ");
+      echo "<script>
+      location.href='tabla.php';
+      </script>";
+   }else{
+      echo "<script>
+      alert('El registro no pudo ser eliminado');
+      location.href='tabla.php';
+      </script>";
+   }
 ?>
